requirements/dev.txt
-r base.txt

# Форматирование и линтинг
black==24.10.0
flake8==7.1.1
flake8-bugbear==24.10.31
flake8-docstrings==1.7.0
flake8-import-order==0.18.2
pycodestyle==2.12.1
pydocstyle==6.3.0
pyflakes==3.2.0
mccabe==0.7.0


# Тесты
pytest==8.3.4
pytest-django==4.9.0
pytest-mock==3.14.0
pytest-cov==6.0.0
coverage==7.6.9
factory_boy==3.3.1
Faker==33.3.1
freezegun==1.5.1
